Splitting out account pages & adding user nav bar
