+++ lib/glimesh_web.ex
@@ -35,7 +35,14 @@ defmodule GlimeshWeb do
 
       # Import convenience functions from controllers
       import Phoenix.Controller,
-        only: [get_flash: 1, get_flash: 2, view_module: 1, view_template: 1]
+        only: [
+          get_flash: 1,
+          get_flash: 2,
+          view_module: 1,
+          view_template: 1,
+          action_name: 1,
+          controller_module: 1
+        ]
 
       # Include shared imports and aliases for views
       unquote(view_helpers())
