+++ lib/glimesh_web/templates/layout/app.html.eex
@@ -7,4 +7,3 @@
 <% end %>
 
 <%= @inner_content %>
-
