+++ lib/glimesh_web/templates/layout/user-sidebar.html.eex
@@ -1,0 +1,34 @@
+
+<div class="row">
+    <div class="col-2 user-sidebar">
+        <div class="list-group list-group-flush">
+            <%= link class: "list-group-item list-group-item-action " <> active_user_profile_path(@conn), to: Routes.user_settings_path(@conn, :profile) do %>
+                <i class="fas fa-user fa-fw"></i> <%= dgettext("user", "Profile") %>
+            <% end %>
+            <%= link class: "list-group-item list-group-item-action " <> active_user_stream_path(@conn), to: Routes.user_settings_path(@conn, :stream) do %>
+                <i class="fas fa-video fa-fw"></i> <%= dgettext("user", "Stream") %>
+            <% end %>
+            <%= link class: "list-group-item list-group-item-action " <> active_user_payments_path(@conn), to: Routes.user_payments_path(@conn, :index) do %>
+                <i class="fas fa-money-bill fa-fw"></i> <%= dgettext("user", "Payments") %>
+            <% end %>
+            <%= link class: "list-group-item list-group-item-action " <> active_user_settings_path(@conn), to: Routes.user_settings_path(@conn, :settings) do %>
+                <i class="fas fa-cog fa-fw"></i> <%= dgettext("user", "Settings") %>
+            <% end %>
+            <%= link class: "list-group-item list-group-item-action " <> active_user_security_path(@conn), to: Routes.user_security_path(@conn, :index) do %>
+                <i class="fas fa-lock fa-fw"></i> <%= dgettext("user", "Security") %>
+            <% end %>
+        </div>
+    </div>
+    <div class="col-md-10">
+
+        <%= if get_flash(@conn, :info) do %>
+        <p class="alert alert-info" role="alert"><%= get_flash(@conn, :info) %></p>
+        <% end %>
+        <%= if get_flash(@conn, :error) do %>
+        <p class="alert alert-danger" role="alert"><%= get_flash(@conn, :error) %></p>
+        <% end %>
+
+        <%= @inner_content %>
+
+    </div>
+</div>
