+++ lib/glimesh_web/templates/user_settings/settings.html.eex
@@ -1,0 +1,1 @@
+<h2>If you could settings, your settings would go here.</h2>
