+++ lib/glimesh_web/templates/user_settings/stream.html.eex
@@ -1,0 +1,1 @@
+<h2>If you could stream, your settings would go here.</h2>
