+++ lib/glimesh_web/views/user_security_view.ex
@@ -1,0 +1,3 @@
+defmodule GlimeshWeb.UserSecurityView do
+  use GlimeshWeb, :view
+end
