+++ test/glimesh_web/controllers/user_session_controller_test.exs
@@ -36,7 +36,7 @@ defmodule GlimeshWeb.UserSessionControllerTest do
       conn = get(conn, "/")
       response = html_response(conn, 200)
       assert response =~ user.username
-      assert response =~ "\nMy Profile</a>"
+      assert response =~ "\nSettings</a>"
       assert response =~ "\nSign Out</a>"
     end
